Extract Out PreCommit Message Methods to Utils
Extracted PreCommit context message processor methods to Utils module.
This is a step closer to make these PreCommit hooks usable on other context.
